fixes to task1
small but important and helpul :D
