+++ src/components/HelloFour.jsx
@@ -1,0 +1,9 @@
+import React, { PureComponent } from 'react';
+
+class HelloFour extends PureComponent {
+  render() {
+    return <h1>Hello from pure component</h1>;
+  }
+}
+
+export default HelloFour;
