+++ src/components/HelloOne.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+const HelloOne = () =>
+  React.createElement('h1', {}, 'Hello from create Element');
+
+export default HelloOne;
