+++ src/components/HelloThree.jsx
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class HelloThree extends Component {
+  render() {
+    return <h1>Hello from component</h1>;
+  }
+}
+
+export default HelloThree;
