+++ src/components/HelloTwo.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const HelloTwo = () => <h1>Hello from functional component</h1>;
+
+export default HelloTwo;
